Classes Added to manifest and back buttons added to registration and bookmarks
I added all classes to the manifest as well as added a back button to the registration page. I also added the back button to the bookmarks activity with logic to go back to the main menu. I cleared the login activity as it was doing the same thing as the main activity.